libopeniscsiusr: use realloc instead of reallocarray
realloc is similar to reallocarray, except it doesn't check for integer
overflow when computing num * size.

On 32-bit systems, integer overflow may not happen because the old array
is likely to be smaller than 4G and the new array is smaller than the
old array. On 64-bit systems, integer overflow can never happen because
num is less than 4G and size is always 8.

This fixes build errors when using uClibc.

Signed-off-by: TIAN Yuanhao <tianyuanhao3@163.com>