Build: make iSNS and/or SLP support optional
The idea is to make the iscsid and iscsiadm commands require less
external libraries if discovery via one of these two methods
is not required. This will help where a minimal open-iscsi build
is required.

The Node DB records are supposed to be unmodified by this change,
so that build with or without support can read the same records.