Make UID the default auth type for IPCs
Modern security practices include checking the originating UID
of an IPC. But not the username, as there are many valid scenarios
where a user database may not exist or the root account may not be
explicitly named "root".

The mode to check UID and username is still functional, but is now
on the path to deprecation. Related function names and enums now
include the term "legacy" to indicate this deprecated status.

The 'iscsid.ipc_auth_uid' setting in iscsid.config now defaults to
'Yes', rather than introducing a new key. Users sometimes keep old
config files when upgrading packages, and this ensures behavior
remains the same if this setting was explicitly defined in an old
config.

Signed-off-by: Eric Mackay <eric.mackay@oracle.com>
Reviewed-by: Mike Christie <michael.christie@oracle.com>