Remove unused SLP discovery code
Support for SLP Discovery of targets was never full implemented,
and likely won't be any time soon by this project, so remove
the code that was designed to support it.

Note that this removal does not include the code for
NFS SLP, which is a different beast, and left unchanged.