Fix libopeniscsiusr licensing
The library was licensed GLPv3, but LGPLv3 is more
appropriate for a library used widely by others.

The change is done with permission of the library
author, after some discussion.

This will effectively upgrade the license, in
practice, for open-iscsi, to GPLv3, without
having to actually change the higher-level license.