When writing DB record without a tpgt, take tpgt from existing record (#409)
* idmb_rec_write, check for tpgt first

Factor out the check for a tpgt to a single place, before going crazy on
the rec files.  Makes flow of this function easier to follow, and preps
for splitting it up.

Signed-off-by: Chris Leech <cleech@redhat.com>

* idbm_rec_write, seperate old and new style writes

Duplicates a small bit of code, but easier to understand and extened.

Signed-off-by: Chris Leech <cleech@redhat.com>

* idbm_rec_write, pick tpgt from existing record

On a static add (-m node -o new) without a user specified tpgt, looks
for existing new style records with tpgt before creating an old style
record without.  If one exists, take the tpgt from it an write an
updated new style record instead.

Signed-off-by: Chris Leech <cleech@redhat.com>

---------

Signed-off-by: Chris Leech <cleech@redhat.com>