Build: make open-isns usage optional. (#407)
* Build: make open-isns usage optional.

In some cases open-isns may not be needed or may not
be available. Make building without it optional, using
meson. The depricated Makefile build is not effected.

* Build: ensure "make" enables iSNS support

Even though "make" is deprecated, it's still here,
so make sure it enables iSNS support, since that is what
it has always done. If somebody wants to make iSNS
optional for "make", they can do so.

* Build: clean up how iSNS library is used.

The functionality stays the same, but the meson.build
file is a little easier to understand this way.

Suggested by: Eli Schwartz

* build: clean up comments for isns lib detection