remove unicode symbol from iscsid.conf (#423)
Trivial fixup.  I ran into a test script that failed becuase iscsid.conf
wasn't ASCII.  I like UTF-8, but none of this code handles it. And while
unicode bytes in a comment shouldn't be the end of the world, it's just
a lot clearer to remove the single offending symbol.

Signed-off-by: Chris Leech <cleech@redhat.com>