iscsiuio: rename logging macros using ILOG_*()
This change is in preperation for using syslog
for logging, should we decide to do that, since
syslog uses LOG_*() macro names. So rename ours
to ILOG_*(), to be less confusing.

Also, while modifying the logging lines, go
ahead and fix formatting strings to be on
one line, if possible.

No functional change.