iscsiuio: use syslog for logging
This is a multi-threaded daemon, so rolling
our own logging was difficult, particularly
with respect to locking, which tended to get
deadlocked when shutting down. Just use syslog,
since it's engineered for that.

Note that for foreground mode, we now print
just to stdout, or stderr for error messages.
For background mode, we now use syslog().