Iscsiuio fixes v2 (#427)
* iscsiuio: when handling signals log all of them

Signed-off-by: David Bond <dbond@suse.com>

* iscsiuio: fix formatting '%D' -> '%d'

Fix a LOG_WARN() formatting error.

Signed-off-by: David Bond <dbond@suse.com>

* iscsiuio: fix mispelling: assoicate* -> associate*

Fix several mispellings, some in function names,
most in comments. No functional change.

Signed-off-by: David Bond <dbond@suse.com>

* iscsiuio: change parse_iface() to non-cancelable

This is because this thread holds the nic_list_mutex for
quite a while, so this avoids the mutex being orphanced.

Signed-off-by: David Bond <dbond@suse.com>

* iscsiuio: fix possible logging dealocks

The logging thread can be cancelled while holding
a logging mutex, so ensure that doesn't happen.

Testing shows this eliminated some hangs from mutex deadlock.

Signed-off-by: David Bond <dbond@suse.com>

* iscsiuio: release waiting thread NL when cancelling

This can prevent a mutex deadlock.

Signed-off-by: David Bond <dbond@suse.com>

* iscsiuio: change NL socket receive to have a timeout

This avoids a possible deadlock situations by making the
receive from the netlink socket time out after 1 second
if no data is received, at which point it loops and tries
again. Logging is also slightly modified to avoid a
storm of timeout messages.

Testing showed many successful repititions of service
start/stop without error with this change.

Signed-off-by: David Bond <dbond@suse.com>

* iscsiuio: signal success on exit, and close NL socket

Two small cleanups which help shutdown more cleanly. First,
exit with success if we are sucessful. Second, close (with
no-linger) the netlink socket when we are done with it.

* iscsiuio: remove unused logging statistics

These were not being used, so just remove them.

* iscsiuio: stop logging to file and stdout

This was redundant in the name of being more helpful,
but doubled the load on the daemon for some types of
logging, since it wrote to and flushed both the file
pointer and stdout for each debug message.

* iscsiuio: fix IPv4 address parsing for qedi

When scanning an IPv4 IP address for qedi adapters,
we expect 6 tuples, but check for one, printing an
incorrect error message in every case:

>	NIC-NAME: Could not parse mac_addr

Also, improve the error message while here.

Found-by: David Bond <DBond@suse.com>

* iscsiuio: rename logging macros using ILOG_*()

This change is in preperation for using syslog
for logging, should we decide to do that, since
syslog uses LOG_*() macro names. So rename ours
to ILOG_*(), to be less confusing.

Also, while modifying the logging lines, go
ahead and fix formatting strings to be on
one line, if possible.

No functional change.

* iscsiuio: more logging cleanup

Just removing some extra spaces, where not needed

* iscsiuio: use syslog for logging

This is a multi-threaded daemon, so rolling
our own logging was difficult, particularly
with respect to locking, which tended to get
deadlocked when shutting down. Just use syslog,
since it's engineered for that.

Note that for foreground mode, we now print
just to stdout, or stderr for error messages.
For background mode, we now use syslog().

* iscsiuio: add comment about using SO_LINGER

Has no functional effect.

---------

Signed-off-by: David Bond <dbond@suse.com>