iscsiuio: manage threads better at shutdown
This commit adds ability for iscsiuio to count
threads started, and to wait for them at shutdown,
up to 10 seconds maximum. It also adds a few
messages during that shutdown time if threads are
slow to shut down.

Testing showed that 100 iterations of stop/start
of the iscsiuio service completed without error
with these changes.

Signed-off-by: David Bond <dbond@suse.com>