iscsid: set host workqueue priority correctly
The code was scanning through /proc for processes with the
workqueue name (iscsi_q_HOSTNUM) in the process name, and that
didn't always work in newer kernels, where only the current
workqueue being used is part of the kernel worker thread name.
In addition, this method was slow. This often failed, printing
a message like "Could not set sessionN priority. ..." in the
system journal.

Instead, write the the sysfs workqueue "nice" attribute, using
the same value that we used when previously calling setpriority().
Also, add some debugging that can be enabled.