iscsid: Rescan devices on relogin
If a device goes to offline/transport-offline when we are disconnected
the device's size can be set to zero if some other app forces it to get
rescanned. Normally in the kernel when a device is set from offline to
running again, it will rescan the device. However, because during a
relogin the iscsi layer sets the state from offline/transport-offline
via iscsi_unblock_session we don't hit that rescan. We actually don't
want to hit the kernel's rescan right now, because it can block and
that would block all iscsid operations.

This has us perform a rescan operation after a relogin from the scan
process.