Fix authmethod check by printing a warning message when CHAP used and authmethod=None (#443)
* Fix indentation in usr/initiator.c

I just happened upon this code, and had to fix
the indentation-by-spaces. No functional change.

* Warn users of login authorization failure.

Currently, during session establishment, CHAP credentials
are sent, if needed, but any errors doing so are ignored.

Start paying attention to authorization failure. For now, we
will just print a warning message that this behavior is being
deprecated. Some day this should actually be fixed.

* Fix wording on deprecated warning for auth failures.

* Warn when authmethod=None but CHAP credentials set

Print a warning message that we are using CHAP even
though authmethod is unset, or set to None, and that
this is being deprecated "some day".

* Updated iscsid.conf on how "authmethod" attribute used

Just adding comments about authmethod being mostly ignored.
No functional change.

* Add comment for deprecated authmethod handling.

Just added a "FIXNE" comment, to be consistent.