Modify how workqueue priority is set (#445)
The open-iscsi code has been scanning the process table when
it creates a new session (and hence a new host), in order
to raise the priority of the workqueue transmit thread, to
avoid a possible priority inversion with kernel threads like
bdflush.

In modern kernels the workqueue transmit thread can be passive,
meaning that it will not show up in the process table unless it
is actively doing work. This means that we can no longer count
on scanning the process table. But the good news is that those
modern workqueue threads are designed differently and will not
get deadlocked at their normal priority.

For now, we just optimize the code so that if the requested
priority is zero, we do nothing, and we change the default value
to zero, as well.

In the future the process-table-scanning code should be removed.