Better handle multiple iscsiadm commands
The DB locking was racing, at times, since it did something
like:

    if (access() != 0)
	    if (mkdir() != 0)
		    error

But when multiple processes are running, it is possible for
a process to have access() fail, but then mkdir also fail, since
another process has created the directory between the check and
the attempt at creation.

This essentially makes EEXIST of the directory ok, with some
other cleanups of the locking code (in two places).