systemd: Skip units when running in container
iscsi might be installed in an image that can be booted both as a
container and as a virtual machine (or bare metal). Let's condition
out the units when running in a container as iscsi can't reliably do
its job in a container anyways.

Without this commit, iscsi may fail to start when booting an image with
iscsi installed in a systemd-nspawn container.