Fix firmware targets startup to always be "onboot"
Commit 9d99b50a88ef ("Set default 'startup' to 'onboot' for FW nodes")
added code to always set "onboot" for the startup mode for firmware
targets, but it did so only for when targets database entries are
updated, not when they are created.

This commit fixes that by moving the code that sets startup entries
to "onboot" so that it happens for all firmware DB entries.

Fixes: 9d99b50a88ef ("Set default 'startup' to 'onboot' for FW nodes")