testing: update testing discovery and fix bug
The tests used to require the node to already be in
the database before the tests were run, but that was not
clear, so run discovery if needed, and then fail if the
node is not in the database.

Also, fix a bug where dprint was called incorrectly, and
clean up some of the comments.