More testing cleanup, and fix dprint test usage (#491)
* Clean up the git ignore file

No need for some entries now that we use meson/ninja

* Test cleanup: remove ol regression.sh script

And update README to remove reference to it.

* Testing cleanup: remove old regression data

The regression.dat file was used by regression.sh, which is now
removed. And the original testing README is also removed.

* testing: update testing discovery and fix bug

The tests used to require the node to already be in
the database before the tests were run, but that was not
clear, so run discovery if needed, and then fail if the
node is not in the database.

Also, fix a bug where dprint was called incorrectly, and
clean up some of the comments.

* testing: update README and TODO

Add README info about the node db, and add a to do item
for cleaning up the node DB when done.