iscsid: Rate limit session reopen log messages
When the iscsi targets disappear due to any reason like ports down or the
network issues or any other reason, existing sessions will be retried by
the iscsi daemon and each retry will have a 3 seconds/DefaultTime2Wait gap.
So, the session reopen logs are logged into logging system for every 3
seconds and due to this the logging system is overwhelmed. To avoid this
overwhelming of logging, a config parameter is provided in iscsid.conf file
in this feature using which a user can tune the interval between the
subsequent session reopen log messages. By default the value of this config
option in iscsid.conf file (node.session.reopen_log_freq) is 1 and is
commented out currently.
If user wants to change this value, he can uncomment this option and set
the value to a desirable value like 2 or 3 or 4 and restart iscsid. If the
value for this config value is 4, then these session reopen logs will be
logged for every 12 seconds (4 * 3 seconds = 12 seconds).
This feature works only if the default log level in iscsid.service file
is set to <= 2

Also changed the log level of a debug print in usr/event_poll.c in the function
event_loop(), from 1 to 4 to avoid the overwhelming the logging system
as these prints are printed for every session reopen.
Also changed the log_error() to log_debug_rate_limited() in the function
iscsi_io_tcp_poll() in file usr/io.c only if this feature is enabled.

Signed-off-by: Gulam Mohamed <gulam.mohamed@oracle.com>