iscsid: Rate limit session reopen log messages
When the iscsi targets disappear due to any reason like ports down or the
network issues or any other reason, existing sessions will be retried by
the iscsi daemon and each retry will have a 3 seconds gap by default. So,
the session reopen logs are logged into /var/log/messages for every 3
seconds and due to this the messages file is blown out. To avoid this blow
out of messages file, a config parameter is provided in iscsid.conf file
in this feature using which a user can tune the interval between the
subsequent session reopen log messages By default the value of this config
option in iscsid.conf file (node.session.reopen_log_freq) is 1 and is
commented out currently.
If user wants to change this value, he can uncomment this option and set
the value to a desirable value like 2 or 3 or 4 and restart iscsid. If the
value for this config value is 4, then these session reopen logs will be
logged for every 12 seconds (4 * 3 seconds = 12 seconds).
This feature works only if the default log level in iscsid.service file
is set to <= 2

Signed-off-by: Gulam Mohamed <gulam.mohamed@oracle.com>