Improve iscsiadm command line parsing messages
Currently, if you pass in an unknown "mode" to the
iscsiadm command, you get a message like:

> # iscsiadm -m zzz
> iscsiadm: mode -1 is not yet supported

And this is not informative or correct. Also, if you pass
in a valid mode but an invalid option, you get:

> # iscsiadm -m session -z
> iscsiadm: invalid option -- 'z'
> iscsiadm: session mode: option '-?' is not allowed/supported

Again, the "-?" is not correct or useful. Both of these are
improved with this commit.