iscsid: Rate limit session reopen log messages (#492)
* iscsid: Rate limit session reopen log messages

When the iscsi targets disappear due to any reason like ports down or the
network issues or any other reason, existing sessions will be retried by
the iscsi daemon and each retry will have a 3 seconds/DefaultTime2Wait gap.
So, the session reopen logs are logged into logging system for every 3
seconds and due to this the logging system is overwhelmed. To avoid this
overwhelming of logging, a config parameter is provided in iscsid.conf file
in this feature using which a user can tune the interval between the
subsequent session reopen log messages. By default the value of this config
option in iscsid.conf file (node.session.reopen_log_freq) is 1 and is
commented out currently.
If user wants to change this value, he can uncomment this option and set
the value to a desirable value like 2 or 3 or 4 and restart iscsid. If the
value for this config value is 4, then these session reopen logs will be
logged for every 12 seconds (4 * 3 seconds = 12 seconds).
This feature works only if the default log level in iscsid.service file
is set to <= 2

Also changed the log level of a debug print in usr/event_poll.c in the function
event_loop(), from 1 to 4 to avoid the overwhelming the logging system
as these prints are printed for every session reopen.
Also changed the log_error() to log_debug_rate_limited() in the function
iscsi_io_tcp_poll() in file usr/io.c only if this feature is enabled.

Signed-off-by: Gulam Mohamed <gulam.mohamed@oracle.com>

* Addressed the following comments:

1. Clarified/Modified the following in the comments and description:
    a. Why this feature works only if the default log level in
       iscsid.service file is set to <= 2
    b. Why we use int64_t for reopen_log_freq field in struct
       iscsi_session_idbm
    c. Why the log level of log_debug() in event_loop() is raised
2. Renamed the logging function to be more generic to log_debug_rate_limited()
3. Modified some of the comments:
    a. Changed the wording "/var/log/messages" to be generic like
       "logging system"
    b. Provided more clarity in the comment for the define
       DEF_SESSION_REOPEN_MAX
4. Using defines now wherever the default value is assigned to
   reopen_log_freq
5. Changed the logic in the function log_debug_rate_limited()to take care
   to not print any message when the logging is disabled
6. Rolled back to log level 4 for log_debug() in the function
   actor_delete() as suggested by Mike and will make it to log level 1 in a
   separate patch
7. Changed the logic in __session_conn_create() to allow reopen_log_freq to
   be 0 to just turn it off
8. Also changed the logic in iscsi_io_tcp_poll() to allow the
   log_debug_rate_limited() only if reopen_log_freq is set

* As per Mike's suggestion for short term, did the following changes:

1. Changed the logging function name from "log_debug_rate_limited()" to
   "conn_log_connect()"
2. Apart from using this logging function just for session reopen messages,
   now this function is used for messages related to the establishing the
   initial tcp connection also. As of now, only log_debug() messages
   with log_level 1 and also the log_info() have been changed in these paths
3. The config option is changed from "reopen_log_freq" to
   "conn_reopen_log_freq"
4. The field "reopen_log_freq" in "struct session_rec" and "struct
   iscsi_session_idbm" changed to "conn_reopen_log_freq"
5. The name of #define for default value of log frequency changed from
   DEF_SESSION_REOPEN_LOG_FREQ to DEF_CONN_REOPEN_LOG_FREQ

Signed-off-by: Gulam Mohamed <gulam.mohamed@oracle.com>

---------

Signed-off-by: Gulam Mohamed <gulam.mohamed@oracle.com>