Fix issue with IPv6 adapter interfaces (#508)
When populating an open-iscsi interface file, using
iscsiadm firmware-mode discovery, for some offload
cards, later use of that interface file requires
the code to guess the whether the interface file is
for IPv4 of IPv6, particularly when the IP address
is has been configured as "default".

Fix this by adding a check in the "what kind of
interface file is this" code for the "default"
IP address.

Suggested-by: David Bond <dbond@suse.com>
Tested-by: David Bond <dbond@suse.com>

Signed-off-by: Lee Duncan <lduncan@suse.com>