iscsi_net_util: avoid copying NULL pointers with strlcpy() (#515)
* iscsi_net_util: avoid copying NULL pointers with strlcpy()

net_setup_netdev_ipv4()/net_setup_netdev_ipv6() inittializes physdev =
NULL and passes it to find_vlan_dev(), then strlcpy() copies from
physdev.This is undefined behavior and should not occur.

And the parameter passed in is only used by strlcpy(), with no other
usages, so the strlcpy() can be directly removed.

Signed-off-by: Wu Guanghao <wuguanghao3@huawei.com>

* iscsi_net_util: find_vlan_dev() remove unused paramter

The parameter netdev is not being used and can be removed.

Signed-off-by: Wu Guanghao <wuguanghao3@huawei.com>

---------

Signed-off-by: Wu Guanghao <wuguanghao3@huawei.com>