Infer interface on link-local target discovery
When a discovery probe is sent to a link-local address, the server can
reply with link-local TargetAddress entries. In order for those to work,
they need to be augmented with an interface qualifier.

This change deals with that need by appending the interface qualifier
used in the discovery probe, if available.

Closes #400.